feat(file): 添加图像解析状态和来源字段
- 在 Image 模型中添加 period 字段用于存储解剖期限
- 在 ImageConstant 中添加图像解析状态和来源的常量
- 修改 ImageServiceImpl 中的文件名解析逻辑，增加解剖期限字段
- 更新 FileController 和 FileManagerController 中的相关接口

+++ src/main/java/cn/staitech/file/controller/FileController.java
@@ -71,7 +71,7 @@ public class FileController {
 			}
 
 			// step3:校验文件是否重复：校验文件名
-			int count = imageService.count(Wrappers.<Image>lambdaQuery().eq(Image::getImageName, fileInformation.getImageName()));
+			long count = imageService.count(Wrappers.<Image>lambdaQuery().eq(Image::getImageName, fileInformation.getImageName()));
 			if (check && count>0) {
 				return R.fail(ImageConstant.IMAGE_EXISTS);
 			}
